refactor(logo): :recycle: consolidate logo component and update imports
- Remove landing-specific logo component and story
- Move shared logo functionality to shared component
- Update imports in header components
- Remove unused public routes
